Moved device specific producer logic to its own abstract class
